+++ Areas/Management/ViewModel/ManagementVm.cs
@@ -21,22 +21,22 @@ namespace PurityERP.Areas.Management.ViewModel
         public DateTime PurchaseDate { get; set; }
         
         public int PurchaseQuantity { get; set; }
-        public int UnitPrice { get; set; }
+        public decimal UnitPrice { get; set; }
 
-        public int RemainingQty { get; set; }
+        public decimal RemainingQty { get; set; }
         public int InventoryItem { get; set; }
         public DateTime SystemDate { get; set; }
         public string ProductTittle { get; set; }
-        public int ProductQuantity { get; set; }
-        public int InventoryQuantity { get; set; }
+        public decimal ProductQuantity { get; set; }
+        public decimal InventoryQuantity { get; set; }
         public string Worker { get; set; }
         public string ProductCode { get; set; }
-        public float PerProductInventoryQuantity { get; set; }
+        public decimal PerProductInventoryQuantity { get; set; }
 
 
-        public int CostingPrice { get; set; }
-        public int SalesPrice { get; set; }
-        public int DiscountRate { get; set; }
+        public decimal CostingPrice { get; set; }
+        public decimal SalesPrice { get; set; }
+        public decimal DiscountRate { get; set; }
         public int InitialProductStockQty { get; set; }
 
         public int pId { get; set; }
@@ -52,11 +52,11 @@ namespace PurityERP.Areas.Management.ViewModel
         public DateTime WorkAsignDate { get; set; }
         public string WorkType { get; set; }
         public float PerUnitCost { get; set; }
-        public int Quantity { get; set; }
+        public decimal Quantity { get; set; }
         public DateTime EDD { get; set; }
         public DateTime ADD { get; set; }
         public int DeliveryQty { get; set; }
-        public int PaidAmount { get; set; }
+        public decimal PaidAmount { get; set; }
         public string WorkStatus { get; set; }
 
     }
